make a component to custom textfeild
make a component to custom textfeild
