+++ bloomi/lib/components/custom_textfeild.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class CustomTextFeild extends StatelessWidget {
+  const CustomTextFeild({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return TextField(
+      decoration: InputDecoration(
+        labelText: "Enter your full name",
+        border: OutlineInputBorder(borderRadius: BorderRadius.circular(10)),
+      ),
+    );
+  }
+}
